DROTH-588 Point assets snapping to link fixed, refactored
